move handlers to their own class
