+++ lib/subscribers/azure-functions/index.js
@@ -6,18 +6,14 @@
 'use strict'
 
 const Subscriber = require('../base')
-const Transaction = require('#agentlib/transaction/index.js')
-const recordWeb = require('#agentlib/metrics/recorders/http.js')
-const backgroundRecorder = require('#agentlib/metrics/recorders/other.js')
-const { Transform } = require('node:stream')
-const { DESTINATIONS: DESTS, TYPES } = Transaction
+const HttpHandler = require('#agentlib/subscribers/azure-functions/http-handler.js')
+const BackgroundHandler = require('#agentlib/subscribers/azure-functions/background-handler.js')
 
 module.exports = class AzureFunctionsSubscriber extends Subscriber {
   constructor({ agent, logger }) {
     super({ agent, logger, channelName: 'nr_generic', packageName: '@azure/functions' })
     this.requireActiveTx = false
     this.coldStart = true
-    this.logHookRegistered = false
   }
 
   handler(data, ctx) {
@@ -41,103 +37,13 @@ module.exports = class AzureFunctionsSubscriber extends Subscriber {
 
     args[1].handler = async function wrappedHandler(...handlerArgs) {
       if (options?.return?.type === 'http') {
-        return self.httpHandler({ thisArg: this, handlerArgs, originalHandler })
+        const handler = new HttpHandler({ subscriber: self })
+        return handler.httpHandler({ thisArg: this, handlerArgs, originalHandler })
       } else {
-        return self.backgroundHandler({ thisArg: this, originalHandler, handlerArgs })
+        const handler = new BackgroundHandler({ subscriber: self })
+        return handler.backgroundHandler({ thisArg: this, originalHandler, handlerArgs })
       }
     }
-  }
-
-  async httpHandler({ thisArg, originalHandler, handlerArgs }) {
-    const ctx = this.agent.tracer.getContext()
-    if (ctx.transaction != null) {
-      // If a transaction already exists, we run the function under that
-      // transaction.
-      return originalHandler.apply(thisArg, handlerArgs)
-    }
-
-    const [request, context] = handlerArgs
-
-    const newCtx = this.#createHttpTransaction(request, context, ctx)
-    const transaction = newCtx.transaction
-    this.addFaasAttributes(transaction, context)
-    const self = this
-    function inContext() {
-      self.#initializeWeb(transaction, request)
-      return originalHandler.apply(this, arguments)
-    }
-    const result = await self.agent.tracer.bindFunction(inContext, newCtx).apply(thisArg, handlerArgs)
-    this.handleColdStart(transaction)
-    transaction.finalizeWeb({ statusCode: result?.status, headers: result?.headers, end: false })
-    if (result?.body instanceof Transform) {
-      result.body.on('close', () => transaction.end())
-    } else {
-      transaction.end()
-    }
-    return result
-  }
-
-  #createHttpTransaction(request, context, currentCtx) {
-    const transaction = new Transaction(this.agent)
-    currentCtx = currentCtx.enterTransaction(transaction)
-
-    transaction.setPartialName(`AzureFunction/${context.functionName}`)
-
-    currentCtx = this.createSegment({
-      name: request.url,
-      recorder: recordWeb,
-      ctx: currentCtx
-    })
-
-    transaction.baseSegment = currentCtx.segment
-
-    return currentCtx
-  }
-
-  #initializeWeb(transaction, request) {
-    const absoluteUrl = request.url
-    const url = new URL(absoluteUrl)
-    const transport = url.protocol === 'https:' ? 'HTTPS' : 'HTTP'
-    const port = url.port || (transport === 'HTTPS' ? 443 : 80)
-    transaction.initializeWeb({ absoluteUrl, method: request.method, port, headers: request.headers, transport })
-  }
-
-  async backgroundHandler({ thisArg, originalHandler, handlerArgs }) {
-    const ctx = this.agent.tracer.getContext()
-    if (ctx.transaction != null) {
-      return originalHandler.apply(thisArg, handlerArgs)
-    }
-    const [, context] = handlerArgs
-
-    const newCtx = this.#createBgTransaction(context, ctx)
-    const transaction = newCtx.transaction
-
-    this.addFaasAttributes(transaction, context)
-    const result = await this.agent.tracer.bindFunction(originalHandler, newCtx).apply(thisArg, handlerArgs)
-    this.handleColdStart(transaction)
-
-    transaction.end()
-    return result
-  }
-
-  #createBgTransaction(context, currentCtx) {
-    const transaction = new Transaction(this.agent)
-    currentCtx = currentCtx.enterTransaction(transaction)
-
-    transaction.type = TYPES.BG
-    transaction.setPartialName(`AzureFunction/${context.functionName}`)
-
-    const triggerType = context.options?.trigger?.type ?? 'unknown'
-    const methodName = triggerType.replace(/Trigger$/, '')
-    currentCtx = this.createSegment({
-      name: `${methodName}-trigger`,
-      recorder: backgroundRecorder,
-      ctx: currentCtx
-    })
-
-    transaction.baseSegment = currentCtx.segment
-
-    return currentCtx
   }
 
   get missingEnvVars() {
@@ -164,92 +70,5 @@ module.exports = class AzureFunctionsSubscriber extends Subscriber {
   get azureFunctionAppName() {
     const { WEBSITE_SITE_NAME } = process.env
     return WEBSITE_SITE_NAME
-  }
-
-  addFaasAttributes(transaction, functionContext) {
-    transaction.trace.attributes.addAttribute(
-      DESTS.TRANS_COMMON,
-      'faas.invocation_id',
-      functionContext.invocationId ?? 'unknown'
-    )
-    transaction.trace.attributes.addAttribute(
-      DESTS.TRANS_COMMON,
-      'faas.name',
-      functionContext.functionName ?? 'unknown'
-    )
-    transaction.trace.attributes.addAttribute(
-      DESTS.TRANS_COMMON,
-      'faas.trigger',
-      this.mapTriggerType(functionContext)
-    )
-    transaction.trace.attributes.addAttribute(
-      DESTS.TRANS_COMMON,
-      'cloud.resource_id',
-      this.buildCloudResourceId(functionContext)
-    )
-  }
-
-  mapTriggerType(functionContext) {
-    const input = functionContext.options?.trigger?.type
-
-    // Input types are found at:
-    // https://github.com/Azure/azure-functions-nodejs-library/blob/138c021/src/trigger.ts
-    // https://learn.microsoft.com/en-us/azure/azure-functions/functions-triggers-bindings?tabs=isolated-process%2Cnode-v4%2Cpython-v2&pivots=programming-language-javascript#supported-bindings
-    switch (input) {
-      case 'httpTrigger': {
-        return 'http'
-      }
-
-      case 'timerTrigger': {
-        return 'timer'
-      }
-
-      case 'blobTrigger':
-      case 'cosmosDBTrigger':
-      case 'daprBindingTrigger':
-      case 'mysqlTrigger':
-      case 'queueTrigger':
-      case 'sqlTrigger': {
-        return 'datasource'
-      }
-
-      case 'daprTopicTrigger':
-      case 'eventGridTrigger':
-      case 'eventHubTrigger':
-      case 'kafkaTrigger':
-      case 'rabbitMQTrigger':
-      case 'redisListTrigger':
-      case 'redisPubSubTrigger':
-      case 'redisStreamTrigger':
-      case 'serviceBusTrigger':
-      case 'signalRTrigger':
-      case 'webPubSubTrigger': {
-        return 'pubsub'
-      }
-
-      default: {
-        return 'other'
-      }
-    }
-  }
-
-  buildCloudResourceId(functionContext) {
-    return [
-      '/subscriptions/',
-      this.subscriptionId,
-      '/resourceGroups/',
-      this.resourceGroup,
-      '/providers/Microsoft.Web/sites/',
-      this.azureFunctionAppName,
-      '/functions/',
-      functionContext.functionName
-    ].join('')
-  }
-
-  handleColdStart(transaction) {
-    if (this.coldStart === true) {
-      transaction.trace.attributes.addAttribute(DESTS.TRANS_COMMON, 'faas.coldStart', true)
-      this.coldStart = false
-    }
   }
 }
